+++ native~/src/Runtime/UnityPrepareRendererResources.cpp
@@ -74,6 +74,7 @@
 
 #include <algorithm>
 #include <array>
+#include <cstddef>
 #include <unordered_map>
 #include <variant>
 
@@ -153,7 +154,7 @@ template <typename TIndex> struct CopyVertexColors {
     uint8_t a;
   };
 
-  bool operator()(AccessorView<nullptr_t>&& invalidView) { return false; }
+  bool operator()(AccessorView<std::nullptr_t>&& invalidView) { return false; }
 
   template <typename TColorView> bool operator()(TColorView&& colorView) {
     if (colorView.status() != AccessorViewStatus::Valid) {